x1plusd: Modularize led animations
